[92] move EventType enum to enums package and update import in EventDateLocation
