+++ dao/src/main/java/greencity/entity/event/EventDateLocation.java
@@ -1,5 +1,6 @@
 package greencity.entity.event;
 
+import greencity.enums.EventType;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
